idrt/algorithm/requirements.txt
parsons
python-dotenv
annoy
pypika